+++ src/models/raster/types.ts
@@ -1,0 +1,3 @@
+import { RasterProductTypes } from './constants';
+
+export type RasterProductTypes = (typeof RasterProductTypes)[keyof typeof RasterProductTypes];